namespace all serializers to V1
