+++ app/serializers/api/v1/participant_validation_serializer.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require "jsonapi/serializer/instrumentation"
+
+module Api
+  module V1
+    class ParticipantValidationSerializer
+      include JSONAPI::Serializer
+      include JSONAPI::Serializer::Instrumentation
+
+      set_id :trn
+      attributes :trn, :qts, :active_alert
+    end
+  end
+end
