--- app/serializers/partnership_csv_serializer.rb
@@ -1,39 +1,0 @@
-# frozen_string_literal: true
-
-require "csv"
-
-class PartnershipCsvSerializer
-  attr_reader :scope
-
-  def initialize(scope)
-    @scope = scope
-  end
-
-  def call
-    CSV.generate do |csv|
-      csv << csv_headers
-
-      scope.each do |record|
-        csv << to_row(record)
-      end
-    end
-  end
-
-private
-
-  def csv_headers
-    %w[
-      urn
-      name
-      delivery_partner
-    ]
-  end
-
-  def to_row(record)
-    [
-      record.school.urn,
-      record.school.name,
-      record.delivery_partner&.name,
-    ]
-  end
-end
